imageVector/requirements.txt
transformers==4.42.3
chromadb==0.5.3
matplotlib==3.9.0
pillow==10.4.0
mysql-connector-python==9.0.0
torch==2.3.1+cu118
torchvision==0.18.1+cu118
torchaudio==2.3.1+cu118